Move process kill endpoint
